Move test_insert_line_string_conversion from test_file_edit_action.py to test_aci_edit.py
